+++ demo/admin/src/products/ProductsGrid.tsx
@@ -166,7 +166,6 @@ function ProductsGrid() {
                                             })),
                                             articleNumbers: input.articleNumbers,
                                             discounts: input.discounts,
-                                            packageDimensions: input.packageDimensions,
                                             statistics: { views: 0 },
                                         },
                                     },
@@ -245,11 +244,6 @@ const productsFragment = gql`
         discounts {
             quantity
             price
-        }
-        packageDimensions {
-            width
-            height
-            depth
         }
     }
 `;