+++ demo/api/src/products/entities/product.entity.ts
@@ -3,8 +3,6 @@ import { CrudField, CrudGenerator, DamImageBlock, DocumentInterface, RootBlockDa
 import {
     BaseEntity,
     Collection,
-    Embeddable,
-    Embedded,
     Entity,
     Enum,
     ManyToMany,
@@ -18,6 +16,7 @@ import {
     types,
 } from "@mikro-orm/core";
 import { Field, ID, InputType, ObjectType, registerEnumType } from "@nestjs/graphql";
+import { Manufacturer } from "@src/products/entities/manufacturer.entity";
 import { IsNumber } from "class-validator";
 import { v4 as uuid } from "uuid";
 
@@ -56,26 +55,6 @@ export class ProductDimensions {
     @IsNumber()
     height: number;
 
-    @Field()
-    @IsNumber()
-    depth: number;
-}
-
-@Embeddable()
-@ObjectType()
-@InputType("ProductPackageDimensionsInput")
-export class ProductPackageDimensions {
-    @Property({ type: types.integer })
-    @Field()
-    @IsNumber()
-    width: number;
-
-    @Property({ type: types.integer })
-    @Field()
-    @IsNumber()
-    height: number;
-
-    @Property({ type: types.integer })
     @Field()
     @IsNumber()
     depth: number;
@@ -130,11 +109,15 @@ export class Product extends BaseEntity<Product, "id"> implements DocumentInterf
     inStock: boolean = true;
 
     @Property({ type: types.decimal, nullable: true })
-    @Field()
+    @Field({ nullable: true })
     @CrudField({
         input: false,
     })
     soldCount?: number;
+
+    @Property({ type: types.date, nullable: true })
+    @Field({ nullable: true })
+    availableSince?: Date;
 
     @Property({ customType: new RootBlockType(DamImageBlock) })
     @Field(() => RootBlockDataScalar(DamImageBlock))
@@ -152,10 +135,6 @@ export class Product extends BaseEntity<Product, "id"> implements DocumentInterf
     @Property({ type: "json", nullable: true })
     @Field(() => ProductDimensions, { nullable: true })
     dimensions?: ProductDimensions = undefined;
-
-    @Embedded(() => ProductPackageDimensions, { nullable: true })
-    @Field(() => ProductPackageDimensions, { nullable: true })
-    packageDimensions?: ProductPackageDimensions = undefined;
 
     @OneToOne(() => ProductStatistics, { inversedBy: "product", owner: true, ref: true, nullable: true })
     @Field(() => ProductStatistics, { nullable: true })
@@ -198,4 +177,7 @@ export class Product extends BaseEntity<Product, "id"> implements DocumentInterf
     @Property({ onUpdate: () => new Date() })
     @Field()
     updatedAt: Date = new Date();
+
+    @ManyToOne(() => Manufacturer, { nullable: true, index: true, ref: true })
+    manufacturer?: Ref<Manufacturer>;
 }