+++ packages/admin/admin/src/index.ts
@@ -121,6 +121,16 @@ export { RouterPromptHandler, SaveAction } from "./router/PromptHandler";
 export { SubRoute, SubRouteIndexRoute, useSubRoutePrefix } from "./router/SubRoute";
 export { RowActionsItem, RowActionsItemProps } from "./rowActions/RowActionsItem";
 export { RowActionsMenu, RowActionsMenuProps } from "./rowActions/RowActionsMenu";
+export {
+    Savable,
+    SavableProps,
+    SaveBoundary,
+    SaveBoundaryApi,
+    SaveBoundaryApiContext,
+    useSavable,
+    useSaveBoundaryApi,
+} from "./saveBoundary/SaveBoundary";
+export { SaveBoundarySaveButton } from "./saveBoundary/SaveBoundarySaveButton";
 export { Selected } from "./Selected";
 export { ISelectionRenderPropArgs, Selection, useSelection } from "./Selection";
 export { ISelectionApi } from "./SelectionApi";
@@ -198,3 +208,5 @@ export { RouterTabsClassKey } from "./tabs/RouterTabs.styles";
 export { Tab, Tabs, TabsProps } from "./tabs/Tabs";
 export { TabsClassKey } from "./tabs/Tabs.styles";
 export { TabScrollButton, TabScrollButtonClassKey, TabScrollButtonProps } from "./tabs/TabScrollButton";
+export { ContentTranslationServiceProvider } from "./translator/ContentTranslationServiceProvider";
+export { useContentTranslationService } from "./translator/useContentTranslationService";