+++ packages/api/cms-api/src/generator/generate-crud.ts
@@ -999,10 +999,13 @@ function generateResolver({ generatorOptions, metadata }: { generatorOptions: Cr
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export async function generateCrud(generatorOptions: CrudGeneratorOptions, metadata: EntityMetadata<any>): Promise<GeneratedFile[]> {
-    generatorOptions.update = generatorOptions.update ?? true;
-    generatorOptions.create = generatorOptions.create ?? true;
-    generatorOptions.delete = generatorOptions.delete ?? true;
+export async function generateCrud(generatorOptionsParam: CrudGeneratorOptions, metadata: EntityMetadata<any>): Promise<GeneratedFile[]> {
+    const generatorOptions = {
+        ...generatorOptionsParam,
+        create: generatorOptionsParam.create ?? true,
+        update: generatorOptionsParam.update ?? true,
+        delete: generatorOptionsParam.delete ?? true,
+    };
 
     const generatedFiles: GeneratedFile[] = [];
 
@@ -1066,5 +1069,8 @@ export async function generateCrud(generatorOptions: CrudGeneratorOptions, metad
         return generatedFiles;
     }
 
-    return [...(await generateCrudInput(generatorOptions, metadata)), ...(await generateCrudResolver())];
+    const crudInput = await generateCrudInput(generatorOptions, metadata);
+    const crudResolver = await generateCrudResolver();
+
+    return generatorOptions.create || generatorOptions.update ? [...crudInput, ...crudResolver] : [...crudResolver];
 }