+++ packages/api/cms-api/src/page-tree/page-tree.constants.ts
@@ -5,3 +5,5 @@ export const PAGE_TREE_ENTITY = "PageTreeNode";
 export const PAGE_TREE_CONFIG = "PageTreeConfig";
 
 export const defaultReservedPaths = ["/admin", "/preview"];
+
+export const SITE_PREVIEW_CONFIG = "SitePreviewConfig";