+++ storybook/src/admin/mui/Dialog.tsx
@@ -14,10 +14,8 @@ type DialogSizeOptions = {
 
 const dialogSizeOptions: DialogSizeOptions = {
     "XS (350px)": "xs",
-    "SM (600px)": "sm",
+    "SM (680px)": "sm",
     "MD (1024px)": "md",
-    "LG (1280px)": "lg",
-    "XL (1920px)": "xl",
     FullWidth: "fullWidth",
 };