+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @comet/eslint-config
 
+## 6.2.1
+
+### Patch Changes
+
+-   @comet/eslint-plugin@6.2.1
+
+## 6.2.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@6.2.0
+
+## 6.1.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@6.1.0
+
 ## 6.0.0
 
 ### Major Changes